refactor: 모집공고 생성, 수정 페이지 layout 적용 (SP-633)
